bot/requirements.txt
aiogram~=3.10.0
motor~=3.5.1
python-dotenv~=1.0.1